feat: update database schema and entities to use BigInt for numerical fields
- Changed numerical fields in various Prisma schema files from Int to BigInt for better handling of large values.
- Updated the InviteLink model to use BigInt for maxUses and countJoined.
- Modified the App model to use BigInt for price.
- Adjusted Feedback, Comment, Template, and other models to use BigInt for rating-related fields.
- Updated the TemplateProfile model to change apps from Json[] to Json.
- Changed the Swimlane and SwimlaneItem models to use BigInt for index and order fields respectively.
- Updated the Prisma datasource provider from postgresql to cockroachdb.
- Added dotenv and drizzle-kit dependencies to package.json for environment variable management and database configuration.
- Created a drizzle.config.ts file for drizzle-kit configuration.
- Updated various Rust entity files to reflect changes in the database schema, ensuring consistency with the new BigInt types.
- Adjusted API routes to handle the new BigInt types for ratings and other numerical fields.

+++ packages/api/src/entity/llm_usage_tracking.rs
@@ -11,9 +11,9 @@ pub struct Model {
     #[sea_orm(column_name = "modelId", column_type = "Text")]
     pub model_id: String,
     #[sea_orm(column_name = "tokenIn")]
-    pub token_in: i32,
+    pub token_in: i64,
     #[sea_orm(column_name = "tokenOut")]
-    pub token_out: i32,
+    pub token_out: i64,
     #[sea_orm(column_type = "Double", nullable)]
     pub latency: Option<f64>,
     #[sea_orm(column_name = "userId", column_type = "Text", nullable)]
